Use constant criteria for likelihood underflow
This reverts changes

7ca78b0
2017-09-01
Change 'too small' criteria

and

0f50c59
2017-04-01
Change small value to powers of 1e-10

These changes caused various models (e.g. 1 a process factor analysis,
and e.g. 2 a Psychometrika Monte Carlo simulation) to fail.

What seemed to happen during these failed optimization is the initial
step put the free parameters into a bad space and the optimizer
recovered poorly.

The mechanism was likely due to using too inconsistent of a criteria
for what a 'too small' likelihood or regime probability is.  Now it
is set to be very very close to the double nearest to zero without
underflow.  This is 1e-322; the log of which is -741.44.