Move ggplot2 back to depends
This reverts cdc1b1f for ggplot2.  The package needs to be "Depends"
to restore the previous behavior for autoplot().

Note:
The general rules are

* Packages whose namespace only is needed to load the package using
library(pkgname) must be listed in the ‘Imports’ field and not in the
‘Depends’ field.

* Packages that need to be attached to successfully load the package
using
library(pkgname) must be listed in the ‘Depends’ field, only.

ggplot2 fits under the second bullet above.  We'll need to note this
in our response to reviewers.