Clean up CRAN check issues induced from poorly mering armadillo branch
I probably broke a bunch of this code.  So, whoever wrote it
should fix it or tell me how to fix it.

In the future, it is a good idea to run CRAN check on your code.
It is also a good idea to write tests.  I'm 100% sure the newly
committed armadillo branch code would not pass any reasonable tests.